+++ pyabc/sampler/multicore_evaluation_parallel.py
@@ -64,9 +64,6 @@ class MulticoreEvalParallelSampler(MultiCoreSampler):
         If set to None, the Number of cores is determined according to
         :func:`pyabc.sge.nr_cores_available`.
     """
-    def __init__(self, n_procs=None):
-        super().__init__()
-        self._n_procs = n_procs
 
     @property
     def n_procs(self):
@@ -87,7 +84,7 @@ class MulticoreEvalParallelSampler(MultiCoreSampler):
             Process(target=work,
                     args=(sample_one, simulate_one, accept_one,
                           queue, n_eval, n_particles),
-                    daemon=True)
+                    daemon=self.daemon)
             for _ in range(self.n_procs)
         ]